fix(api): ci_run.number as a big integer — GitHub Actions run ids overflow int32
The first sync of a GitHub app failed on PostgreSQL with a database
error the page showed as a generic failure. Migration 0011 widens the
column; a fresh install gets it right from 0010.

Closes #224

+++ apps/api/app/core/db/migrations/versions/0010_ci.py
@@ -36,7 +36,7 @@ def upgrade() -> None:
         sa.Column("id", sa.String(length=36), nullable=False),
         sa.Column("app_id", sa.String(length=36), nullable=False),
         sa.Column("source", sa.String(length=255), nullable=False),
-        sa.Column("number", sa.Integer(), nullable=False),
+        sa.Column("number", sa.BigInteger(), nullable=False),
         sa.Column("status", sa.String(length=255), nullable=False),
         sa.Column("name", sa.Text(), nullable=True),
         sa.Column("url", sa.Text(), nullable=True),
